[CHORE/VG-316] 명시적 캐스팅 변경

+++ Source/GA6thFinal_Framework/GameScripts/Scripts/TurnSystem/TurnActor/Character/Player/State/PlayerPlayTurnState.cpp
@@ -456,7 +456,7 @@ void PlayerPlayTurnState::PushWeaponAnimation(Battle::EnemyTargetFlag_ destEnemy
             Player& player = GetPlayer();
             Battle()(player, destEnemy);
             
-            int count = weaponAnim->GetOverrideAnimationCount();
+            int count = (int)weaponAnim->GetOverrideAnimationCount();
             if (1 == count)
             {
                 SetAttackEndAnimation();
